added pagination,usersonline indicator,updated encryption for password

+++ admin/includes/view_all_posts.php
@@ -9,7 +9,7 @@ if(isset($_POST['checkBoxArray'])){
                     $bulk_publish_query = mysqli_query($connection,$query);
                 if(!$bulk_publish_query){
                     die("Query Failed" . mysqli_error($connection));
-                }else{ echo "Update Success";}
+                }
                 break;
                 case 'draft';
                     $query = "UPDATE posts SET post_status = 'draft' WHERE post_id = {$checkBoxValue}; ";
@@ -25,6 +25,34 @@ if(isset($_POST['checkBoxArray'])){
                     die("Query Failed" . mysqli_error($connection));
                 }
                 break;
+                case 'clone':
+                    $query = "SELECT * FROM posts WHERE post_id = '{$checkBoxValue}' ";
+                    $select_post_query = mysqli_query($connection, $query);
+                    while($row = mysqli_fetch_array($select_post_query)) {
+                    $post_title = $row['post_title'];
+                    $post_category_id = $row['post_category_id'];
+                    $post_date = $row['post_date'];
+                    $post_author = $row['post_author'];
+                    $post_status = $row['post_status'];
+                    $post_image = $row['post_image'];
+                    $post_tags = $row['post_tags'];
+                    $post_comment_count = $row['post_comment_count'];
+                    $post_content = $row['post_content'];
+                    $query = "INSERT INTO posts(post_category_id, post_title, post_author, post_date, post_image, post_content, post_tags, post_comment_count, post_status) ";
+                    $query .= "VALUES({$post_category_id}, '{$post_title}', '{$post_author}', now(), '{$post_image}', '{$post_content}', '{$post_tags}', '{$post_comment_count}', '{$post_status}') ";
+                    }
+                    $copy_query = mysqli_query($connection, $query);
+                    if(!$copy_query) {
+                        die("QUERY FAILED"  . mysqli_error($connection));
+                    }
+                    break;
+                    case 'reset_views':
+                        
+                        $query = "UPDATE posts SET post_views = 0 WHERE post_id = '{$checkBoxValue}'; ";
+                        $reset_views_query = mysqli_query($connection,$query);
+                        confirmQuery($reset_views_query);
+                        break;
+
                 default;
                 header("Location: posts.php");
         }
@@ -45,6 +73,10 @@ if(isset($_POST['checkBoxArray'])){
                                    <option value="publish">Publish</option>
                                    <option value="draft">Draft</option>
                                    <option value="delete">Delete</option>
+                                   <option value="clone">Clone</option>
+                                   <option value="reset_views">Reset View Count</option>
+
+
                                    
                                </select>                       
                            </div>
@@ -71,9 +103,12 @@ if(isset($_POST['checkBoxArray'])){
                                    <th>Tags</th>
                                    <th>Content</th>
                                    <th>Comments</th>
-                                   <th>Status</th>
+                                   <th>View Count</th>
+                                   <th>Status</th>                             
                                     <th>Edit</th>    
-                                    <th>Delete</th>                                 
+                                    <th>Delete</th>
+                                    <th>Reset View Count</th>                                 
+
                                </tr>
                            </thead>
                        
@@ -94,6 +129,7 @@ if(isset($_POST['checkBoxArray'])){
         $post_tags = $row['post_tags'];
         $post_comment_count = $row['post_comment_count'];
         $post_status = $row['post_status'];
+        $post_views = $row['post_views'];
             
             
             echo "<tr>";?>
@@ -121,9 +157,11 @@ if(isset($_POST['checkBoxArray'])){
             echo "<td>$post_tags</td>";
             echo "<td>$post_content</td>";
             echo "<td>$post_comment_count</td>";
+            echo "<td>$post_views</td>";
             echo "<td>$post_status</td>";
             echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-            echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+            echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this post?');\" href='posts.php?delete={$post_id}'>Delete</a></td>";
+            echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to reset the view count for this post?');\" href='posts.php?reset_views={$post_id}'>Reset Views</a></td>";
             echo "</tr>";
         
             
@@ -148,30 +186,13 @@ if(isset($_GET['delete'])){
     confirmQuery($delete_query);
     header("Location: posts.php");
     }
+
+if(isset($_GET['reset_views'])){
+    $reset_id = $_GET['reset_views'];
+    $query = "UPDATE posts SET post_views = 0 WHERE post_id =". mysqli_real_escape_string($connection,$reset_id); ;
+    $reset_views_query = mysqli_query($connection,$query);
+    confirmQuery($reset_views_query);
+    header("Location: posts.php");
+}
 ?>                  
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-                      
-    
-                       
+
